Added updateCases, createCases, #7
