--- src/AdaptorV2.js
@@ -1,162 +1,0 @@
-/** @module Adaptor */
-import {
-  execute as commonExecute,
-  expandReferences,
-  composeNextState,
-} from 'language-common';
-import HttpAdaptor from 'language-http';
-import { assembleError, tryJson } from './Utils';
-import request from 'request';
-
-/**
- * Execute a sequence of operations.
- * Wraps `language-common/execute`, and prepends initial state for http.
- * @example
- * execute(
- *   create('foo'),
- *   delete('bar')
- * )(state)
- * @function
- * @param {Operations} operations - Operations to be performed.
- * @returns {Operation}
- */
-export function execute(...operations) {
-  const initialState = {
-    references: [],
-    data: null,
-  };
-
-  return state => {
-    return commonExecute(
-      login,
-      ...operations,
-      cleanupState
-    )({ ...initialState, ...state });
-  };
-}
-
-/**
- * Logs in to Primero.
- * @example
- *  login(state)
- * @function
- * @param {State} state - Runtime state.
- * @returns {State}
- */
-function login(state) {
-  const { url, user, password } = state.configuration;
-  const body = JSON.stringify({
-    user: {
-      user_name: user,
-      password,
-    },
-  });
-
-  const params = {
-    method: 'POST',
-    url: `${url}/api/v2/tokens`,
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body,
-  };
-
-  return new Promise((resolve, reject) => {
-    request(params, function (error, response, body) {
-      error = assembleError({ error, response, params });
-      if (error) {
-        reject(error);
-      } else {
-        const resp = tryJson(body);
-        resolve({ ...state, auth: resp });
-      }
-    });
-  });
-}
-
-/**
- * Removes unserializable keys from the state.
- * @example
- *  cleanupState(state)
- * @function
- * @param {State} state
- * @returns {State}
- */
-function cleanupState(state) {
-  delete state.auth;
-  return state;
-}
-
-/**
- * Get cases from Primero
- * @public
- * @example
- *  getCases({
- *    remote: true,
- *    scope: {
- *      transitions_created_at: 'dateRange||17-Mar-2008.17-Mar-2008',
- *      service_response_types: 'list||referral_to_oscar',
- *    },
- *  }, state => {
- *    state.lastFetch =
- *      state.data.map(x => x.updatedAt)
- *      .sort((a, b) => (b-a))
- *    return state
- *  })
- * @function
- * @param {object} query - an object with a query param at minimum.
- * @param {function} callback - (Optional) Callback function
- * @returns {Operation}
- */
-export function getCases(query, callback) {
-  return state => {
-    const { url } = state.configuration;
-    const { token } = state.auth;
-
-    const params = {
-      method: 'GET',
-      url: `${url}/api/v2/cases`,
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-      qs: query,
-    };
-
-    return new Promise((resolve, reject) => {
-      request(params, function (error, response, body) {
-        error = assembleError({ error, response, params });
-        if (error) {
-          reject(error);
-        } else {
-          console.log(
-            `Primero says: '${response.statusCode} ${response.statusMessage}'`
-          );
-          const resp = tryJson(body);
-          console.log(
-            `${resp.data.length} cases retrieved from request: ${JSON.stringify(
-              response.request,
-              null,
-              2
-            )}.`
-          );
-          const nextState = composeNextState(state, resp.data);
-          if (callback) resolve(callback(nextState));
-          resolve(nextState);
-        }
-      });
-    });
-  };
-}
-
-
-export {
-  alterState,
-  dataPath,
-  dataValue,
-  each,
-  field,
-  fields,
-  lastReferenceValue,
-  merge,
-  sourceValue,
-} from 'language-common';
